dialog: move custom dialogs into separate file
